Workable user facing selection of multiple tracks
The user can select check boxes to choose from three tracks: cluster operator, cnct hire, and app developer. The onboarding app will generate or update a project with all the cards whose tags match any of the tracks chosen. For issues that exist in multiple tracks, currently an error message gets generated, which is a situation that will need further refinement; we may not want an error message to show up for a task that got generated during the same action, just because it appears in two tracks. Also needing improvement is the order in which the cards appear.