Added AvailableTracks as a field to the User Model and adjusted handling within controllers and views accordingly.
Tr read in the learning tracks available, a new field was added to the yaml file and read in in config.go. Remaining work could improve on this by parsing the yaml and reading the available tasks from a collection of tags from all the tasks. Additionally, error messages for re-rendering tasks that got created under a previous track need cleaning up.

Overall, a dynamic rendering of available tracks via a check box web from.