Adding track selection method to controller
app.Setup has all the information from the yaml file, including which tracks we tagged each task with. The new method scans eatch task's tags, determines the unique tracks and appends them to a slice of tracks that this particular yaml file provides.

This concludes the main body of dynamic selection and rendering of multiple tracks. Some cleanup remains, as well as a better descriptor for the user-facting checklist.