Changes when error messages display when creating an onboarding project.
Previously, when a task was tagged for multiple learning tracks, the program would attempt to create multiple cards for the same issue, and send an error when the card had been created by another track already. These changes check for cards having been created from another track, and do not attempt to create a card multiple times in the same run. Errors are retained for the case of a subsequent run creating cards over.